% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addAllColumns}
\alias{addAllColumns}
\title{Standardise data columns to match standard template.}
\usage{
addAllColumns(data)
}
\arguments{
\item{data}{data.frame, after being run through convertData}
}
\value{
data.frame
}
\description{
May add or remove columns of data as needed so that all sets have
the same columns.
}

