\name{convertData}
\alias{convertData}
\title{Convert data to desired format, changing units, variable names}
\usage{
  convertData(studyName, data)
}
\arguments{
  \item{studyName}{folder where data is stored}

  \item{data}{existing data frame}
}
\value{
  modified data frame
}
\description{
  Convert data to desired format, changing units, variable
  names
}

