% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mashData}
\alias{mashData}
\title{Adds studies to the central dataset}
\usage{
mashData(studyNames = getStudyNames(), reprocess = TRUE, verbose = FALSE,
  name = "mashup.rds")
}
\arguments{
\item{studyNames}{Character vector containing study names to be added. The default adds names via getStudyNames()}

\item{reprocess}{If TRUE, will reprocess studies even if they already exist in the data directory}

\item{verbose}{If TRUE, print messages to screen, good for isolating problems}
}
\value{
merged list with three parts: data, reference, contact,
   each is a dataframe with all data combined.
}
\description{
Standardises all folders within one main directory (default=data) and combines them into one single dataset via loadStudy()
}

