% Generated by roxygen2 (4.0.0): do not edit by hand
\name{readDataRaw}
\alias{readDataRaw}
\title{Load raw data from studyName}
\usage{
readDataRaw(studyName)
}
\arguments{
\item{studyName}{folder where data is stored}
}
\value{
dataframe with raw data
}
\description{
Load raw data from studyName
}

