% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readNewData}
\alias{readNewData}
\title{Utility function to read/process dataNew.csv for addNewData}
\usage{
readNewData(studyName)
}
\arguments{
\item{studyName}{folder where data is stored}
}
\description{
Utility function to read/process dataNew.csv for addNewData
}

