% Generated by roxygen2 (4.0.2): do not edit by hand
\name{replaceMethod}
\alias{replaceMethod}
\title{Replaces patterns on a given vector}
\usage{
replaceMethod(replacementList, method, replacement)
}
\arguments{
\item{replacementList}{List with pattern as names and respective positions
generated by \code{\link{grepPosition}}}

\item{method}{character vector of methods from dataMatchColumns.csv
(i.e. dataMatchColumns.csv[["method"]])}

\item{replacement}{Replacement string (i.e. new method code)}
}
\value{
modified \code{method} vector
}
\description{
\code{replaceMethod} will replace patterns (names in
\code{replacementList}) in a vector of methods.
}
\seealso{
\code{\link{changeMethodCode}} and \code{\link{grepPosition}}
}

