\name{setUpFiles}
\alias{setUpFiles}
\title{Sets up files for new study to be added to the database}
\usage{
  setUpFiles(newStudy, quiet = FALSE)
}
\arguments{
  \item{newStudy}{Name of the directory that is to be
  added}

  \item{quiet}{If TRUE, don't print messages about
  progress}
}
\value{
  Nothing; directories are created
}
\description{
  Sets up files for new study to be added to the database
}

