% Generated by roxygen2 (4.0.2): do not edit by hand
\name{startDataMashR}
\alias{startDataMashR}
\title{Setup dataMashR}
\usage{
startDataMashR(dir.data = "data", dir.out = "output",
  dir.config = "config", config.files = list(conversions =
  "variableConversion.csv", variables = "variableDefinitions.csv", post =
  "postProcess.R"))
}
\arguments{
\item{dir.data}{Directory where data files are stored. (default="data")}

\item{dir.out}{Directory where to save output (default="output")}

\item{dir.config}{Directory where configuration files are stored (default="config")}

\item{config.files}{Name of configuration files stored within \code{dir.config}. Must be a list with three elements, \code{conversions}, \code{variables}, \code{post}, each conatinig a filename. (default="list(conversions = 'variableConversion.csv', variables = 'variableDefinitions.csv',post = 'postProcess.R')")}
}
\description{
Establshes configuration for dataMashR. Creates object \code{config} within the environment \code{.mashR}, stored within namespace \code{.dataMashR}.
}

