% Generated by roxygen2 (4.0.2): do not edit by hand
\name{validateConfig}
\alias{validateConfig}
\title{Check package configuration files}
\usage{
validateConfig(...)
}
\arguments{
\item{...}{additonal arguments passed through to validate functions}
}
\description{
Check package configuration files
}

