% Generated by roxygen2 (4.0.0): do not edit by hand
\name{validateSetUp}
\alias{validateSetUp}
\title{Validates data files for particular study}
\usage{
validateSetUp(studyName, silent = FALSE)
}
\arguments{
\item{studyName}{name of dircetory to check}

\item{silent}{run silently (default) or print to screen}
}
\description{
Validates data files for particular study
}

