% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convertData}
\alias{convertData}
\title{Convert data to desired format, changing units, variable names}
\usage{
convertData(studyName, data)
}
\arguments{
\item{studyName}{folder where data is stored}

\item{data}{existing data frame}
}
\value{
modified data frame
}
\description{
Convert data to desired format, changing units, variable names
}

