% Generated by roxygen2 (4.0.2): do not edit by hand
\name{data.path}
\alias{data.path}
\title{Path to a directory containing raw data}
\usage{
data.path(studyName, ...)
}
\arguments{
\item{studyName}{name of dataset being requested}

\item{...}{additonal arguments passed through to file.path}
}
\value{
path to particular dataset
}
\description{
Path to a directory containing raw data
}

