\name{manipulateData}
\alias{manipulateData}
\title{Perform arbitrary manipulations to the raw data}
\usage{
  manipulateData(studyName, data)
}
\arguments{
  \item{studyName}{folder where data is stored}

  \item{data}{raw data, as from readDataRaw}
}
\value{
  data.frame, with manipulations carried out
}
\description{
  Perform arbitrary manipulations to the raw data
}

