% Generated by roxygen2 (4.0.0): do not edit by hand
\name{validateConfig}
\alias{validateConfig}
\title{Check package configuration files}
\usage{
validateConfig(silent = FALSE)
}
\arguments{
\item{silent}{run silently (default) or print to screen}
}
\description{
Check package configuration files
}

