% Generated by roxygen2 (4.0.2): do not edit by hand
\name{validateStudy}
\alias{validateStudy}
\title{Validates data files for particular study}
\usage{
validateStudy(studyName, ...)
}
\arguments{
\item{studyName}{name of dircetory to check}

\item{...}{additonal arguments passed through to validate functions}
}
\description{
Validates data files for particular study
}

